Refactor repo logic from DB to separate interface
This will enable the integration of the repo
secret logic later on
