+++ util/db/cluster.go
@@ -3,8 +3,6 @@ package db
 import (
 	"encoding/json"
 	"fmt"
-	"hash/fnv"
-	"net/url"
 	"strconv"
 	"strings"
 	"sync"
@@ -17,10 +15,7 @@ import (
 	apiv1 "k8s.io/api/core/v1"
 	apierr "k8s.io/apimachinery/pkg/api/errors"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/fields"
 	"k8s.io/apimachinery/pkg/watch"
-	informerv1 "k8s.io/client-go/informers/core/v1"
-	"k8s.io/client-go/tools/cache"
 	"k8s.io/utils/pointer"
 
 	"github.com/argoproj/argo-cd/v2/common"
@@ -80,32 +75,29 @@ func (db *db) ListClusters(ctx context.Context) (*appv1.ClusterList, error) {
 
 // CreateCluster creates a cluster
 func (db *db) CreateCluster(ctx context.Context, c *appv1.Cluster) (*appv1.Cluster, error) {
-	secName, err := ServerToSecretName(c.Server)
+	secName, err := URIToSecretName("cluster", c.Server)
 	if err != nil {
 		return nil, err
 	}
+
 	clusterSecret := &apiv1.Secret{
 		ObjectMeta: metav1.ObjectMeta{
 			Name: secName,
-			Labels: map[string]string{
-				common.LabelKeySecretType: common.LabelValueSecretTypeCluster,
-			},
-			Annotations: map[string]string{
-				common.AnnotationKeyManagedBy: common.AnnotationValueManagedByArgoCD,
-			},
 		},
 	}
 
 	if err = clusterToSecret(c, clusterSecret); err != nil {
 		return nil, err
 	}
-	clusterSecret, err = db.kubeclientset.CoreV1().Secrets(db.ns).Create(ctx, clusterSecret, metav1.CreateOptions{})
+
+	err = db.createSecret(ctx, common.LabelValueSecretTypeCluster, clusterSecret)
 	if err != nil {
 		if apierr.IsAlreadyExists(err) {
 			return nil, status.Errorf(codes.AlreadyExists, "cluster %q already exists", c.Server)
 		}
 		return nil, err
 	}
+
 	return secretToCluster(clusterSecret), db.settingsMgr.ResyncInformers()
 }
 
@@ -124,57 +116,42 @@ func (db *db) WatchClusters(ctx context.Context,
 		return err
 	}
 	handleAddEvent(localCls)
-	clusterSecretListOptions := func(options *metav1.ListOptions) {
-		clusterLabelSelector := fields.ParseSelectorOrDie(common.LabelKeySecretType + "=" + common.LabelValueSecretTypeCluster)
-		options.LabelSelector = clusterLabelSelector.String()
-	}
-	clusterEventHandler := cache.ResourceEventHandlerFuncs{
-		AddFunc: func(obj interface{}) {
-			if secretObj, ok := obj.(*apiv1.Secret); ok {
-				cluster := secretToCluster(secretObj)
-				if cluster.Server == common.KubernetesInternalAPIServerAddr {
-					// change local cluster event to modified or deleted, since it cannot be re-added or deleted
-					handleModEvent(localCls, cluster)
-					localCls = cluster
-					return
-				}
-				handleAddEvent(cluster)
-			}
-		},
-		DeleteFunc: func(obj interface{}) {
-			if secretObj, ok := obj.(*apiv1.Secret); ok {
-				if string(secretObj.Data["server"]) == common.KubernetesInternalAPIServerAddr {
-					// change local cluster event to modified or deleted, since it cannot be re-added or deleted
-					handleModEvent(localCls, db.getLocalCluster())
-					localCls = db.getLocalCluster()
-				} else {
-					handleDeleteEvent(string(secretObj.Data["server"]))
-				}
-			}
-		},
-		UpdateFunc: func(oldObj, newObj interface{}) {
-			if oldSecretObj, ok := oldObj.(*apiv1.Secret); ok {
-				if newSecretObj, ok := newObj.(*apiv1.Secret); ok {
-					oldCluster := secretToCluster(oldSecretObj)
-					newCluster := secretToCluster(newSecretObj)
-					if newCluster.Server == common.KubernetesInternalAPIServerAddr {
-						localCls = newCluster
-					}
-					handleModEvent(oldCluster, newCluster)
-				}
-			}
-		},
-	}
-	indexers := cache.Indexers{cache.NamespaceIndex: cache.MetaNamespaceIndexFunc}
-	clusterSecretInformer := informerv1.NewFilteredSecretInformer(db.kubeclientset, db.ns, 3*time.Minute, indexers, clusterSecretListOptions)
-	clusterSecretInformer.AddEventHandler(clusterEventHandler)
-	log.Info("Starting clusterSecretInformer informers")
-	go func() {
-		clusterSecretInformer.Run(ctx.Done())
-		log.Info("clusterSecretInformer cancelled")
-	}()
 
-	<-ctx.Done()
+	db.watchSecrets(
+		ctx,
+		common.LabelValueSecretTypeCluster,
+
+		func(secret *apiv1.Secret) {
+			cluster := secretToCluster(secret)
+			if cluster.Server == common.KubernetesInternalAPIServerAddr {
+				// change local cluster event to modified or deleted, since it cannot be re-added or deleted
+				handleModEvent(localCls, cluster)
+				localCls = cluster
+				return
+			}
+			handleAddEvent(cluster)
+		},
+
+		func(oldSecret *apiv1.Secret, newSecret *apiv1.Secret) {
+			oldCluster := secretToCluster(oldSecret)
+			newCluster := secretToCluster(newSecret)
+			if newCluster.Server == common.KubernetesInternalAPIServerAddr {
+				localCls = newCluster
+			}
+			handleModEvent(oldCluster, newCluster)
+		},
+
+		func(secret *apiv1.Secret) {
+			if string(secret.Data["server"]) == common.KubernetesInternalAPIServerAddr {
+				// change local cluster event to modified or deleted, since it cannot be re-added or deleted
+				handleModEvent(localCls, db.getLocalCluster())
+				localCls = db.getLocalCluster()
+			} else {
+				handleDeleteEvent(string(secret.Data["server"]))
+			}
+		},
+	)
+
 	return err
 }
 
@@ -224,38 +201,19 @@ func (db *db) UpdateCluster(ctx context.Context, c *appv1.Cluster) (*appv1.Clust
 	return secretToCluster(clusterSecret), db.settingsMgr.ResyncInformers()
 }
 
-// Delete deletes a cluster by name
+// DeleteCluster deletes a cluster by name
 func (db *db) DeleteCluster(ctx context.Context, server string) error {
 	secret, err := db.getClusterSecret(server)
 	if err != nil {
 		return err
 	}
 
-	canDelete := secret.Annotations != nil && secret.Annotations[common.AnnotationKeyManagedBy] == common.AnnotationValueManagedByArgoCD
-
-	if canDelete {
-		err = db.kubeclientset.CoreV1().Secrets(db.ns).Delete(ctx, secret.Name, metav1.DeleteOptions{})
-	} else {
-		delete(secret.Labels, common.LabelKeySecretType)
-		_, err = db.kubeclientset.CoreV1().Secrets(db.ns).Update(ctx, secret, metav1.UpdateOptions{})
-	}
+	err = db.deleteSecret(ctx, secret)
 	if err != nil {
 		return err
 	}
-	return db.settingsMgr.ResyncInformers()
-}
 
-// serverToSecretName hashes server address to the secret name using a formula.
-// Part of the server address is incorporated for debugging purposes
-func ServerToSecretName(server string) (string, error) {
-	serverURL, err := url.ParseRequestURI(server)
-	if err != nil {
-		return "", err
-	}
-	h := fnv.New32a()
-	_, _ = h.Write([]byte(server))
-	host := strings.ToLower(strings.Split(serverURL.Host, ":")[0])
-	return fmt.Sprintf("cluster-%s-%v", host, h.Sum32()), nil
+	return db.settingsMgr.ResyncInformers()
 }
 
 // clusterToData converts a cluster object to string data for serialization to a secret
